// src/app/models/model-floor-plan.ts
export interface FloorPlanDetail {
    serialNo?: string;
    propertyDescription?: string;
    propertyDetailsTitle?: string;
    blockName?: string;
    areaName?: string;
    costPerProperty?: number;
    srno?: string;
    compId?: number;
    financialYear?: string;
    projectName?: string;
    sNo?: number;
    status?: string;
    salesPrice?: number;
    fileUpload?: string; // New field to store the uploaded file path   
}